Cleanup the test utils by ensuring proper separation between Maya, Ufe, Usd and general purpose.
